requirements.txt
keystoneauth1
python-keystoneclient
python-swiftclient
